Moved types over to stdint.h
ZZ

+++ include/libgpt.h
@@ -16,23 +16,24 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 #define HEADER_SIZE 512
 #define LBA_SIZE 512
 
 struct GPTHeader{
-  char               signature[8];
-  char               revision[4];
-  unsigned int       headerSize;   //92 in most cases. 512-92 is just 0s
-  unsigned long       crc32;        //different for primary and backup
-  unsigned int       reserved;
-  unsigned long long LBA1, LBA2;   //swapped for backup
-  unsigned long long LBApart, LBApartLast; //different for primary and backup
-  char               GUID[16];
-  unsigned long long LBAstart;     //LBA of partition entries. Different on Pr and Ba
-  unsigned int       numParts;
-  unsigned int       singleSize;
-  unsigned int       crc32Part;
+  char     signature[8];
+  char     revision[4];
+  uint32_t headerSize;   //92 in most cases. 512-92 is just 0s
+  uint32_t crc32;        //different for primary and backup
+  uint32_t reserved;
+  uint64_t LBA1, LBA2;   //swapped for backup
+  uint64_t LBApart, LBApartLast; //different for primary and backup
+  char     GUID[16];
+  uint64_t LBAstart;     //LBA of partition entries. Different on Pr and Ba
+  uint32_t numParts;
+  uint32_t singleSize;
+  uint32_t crc32Part;
 };
 
 
@@ -44,8 +45,8 @@ This function checks both the default and backup GPT headers.
 int isGPT(FILE *deviceFile);
 
 
-unsigned long long getPrimaryHeaderOffset();
-unsigned long long getSecondaryHeaderOffset(FILE *deviceFile);
+uint64_t getPrimaryHeaderOffset();
+uint64_t getSecondaryHeaderOffset(FILE *deviceFile);
 
 
 /*
@@ -54,13 +55,13 @@ Returns 0 otherwise
 */
 int verifyGPT(struct GPTHeader *header);
 
-unsigned long crc32GPT(struct GPTHeader *header);
+uint32_t crc32GPT(struct GPTHeader *header);
 
 
 //Populates the GPTHeader struct
-void readGPT(struct GPTHeader *header, FILE *deviceFile, unsigned long long offset);
+void readGPT(struct GPTHeader *header, FILE *deviceFile, uint64_t offset);
 
-void readCharGPT(char *dstHeader, FILE *deviceFile, unsigned long long offset);
+void readCharGPT(char *dstHeader, FILE *deviceFile, uint64_t offset);
 
 void charToGPTHeader(struct GPTHeader *dst, char *src);
 
@@ -70,14 +71,14 @@ Write header to disk
 This does not alter the partition table
 Returns 0 on clean write, -1 on failure
 */
-int writeCharGPT(char *srcHeader, FILE *deviceFile, unsigned long long offset);
+int writeCharGPT(char *srcHeader, FILE *deviceFile, uint64_t offset);
 
 /*
 Writes the GPT found in the struct GPTHeader to disk
 both primary and backup
 Returns 0 on clean write, -1 on failure
 */
-int writeGPT(struct GPTHeader *header, FILE *deviceFile, unsigned long long offset);
+int writeGPT(struct GPTHeader *header, FILE *deviceFile, uint64_t offset);
 
 void GPTHeaderToChar(char *dst, struct GPTHeader *src);
